+++ foxes/models/turbine_types/calculator_type.py
@@ -1,6 +1,6 @@
 from foxes.core import TurbineType
 
-class DirectDataType(TurbineType):
+class CalculatorType(TurbineType):
     """
     Direct data infusion by a user function.